Change default to named exports in remaining recipes, add missing override tests
