+++ lib/ts/recipe/thirdpartyemailpassword/components/themes/signInAndUp/index.tsx
@@ -29,7 +29,7 @@ import { default as EmailPasswordSignInAndUp } from "../../../../emailpassword/c
 import { SignInAndUpForm as EmailPasswordSignInAndUpForm } from "../../themes/signInAndUp/signInAndUpForm";
 import { SignInAndUpThemeProps as ThirdPartySignInAndUpThemeProps } from "../../../../thirdparty/types";
 import { SignInAndUpThemeProps as EmailPasswordSignInAndUpThemeProps } from "../../../../emailpassword/types";
-import SignInAndUpProvidersForm from "../../../../thirdparty/components/themes/signInAndUp/providersForm";
+import { ProvidersForm } from "../../../../thirdparty/components/themes/signInAndUp/providersForm";
 import { defaultPalette } from "../../../../../styles/styles";
 import { getStyles } from "../styles";
 import { getWindowOrThrow } from "../../../../../utils";
@@ -79,7 +79,7 @@ class SignInAndUpTheme extends React.PureComponent<
                                 recipe={this.props.thirdPartyRecipe}
                                 history={this.props.history}
                                 isEmbedded={true}>
-                                <SignInAndUpProvidersForm
+                                <ProvidersForm
                                     // Seed props. Real props will be given by parent feature.
                                     {...({} as ThirdPartySignInAndUpThemeProps)}
                                 />
